Improve typing around Bitcoin scripts
The concept of a "script" is very widespread in the Bitcoin land. That makes
their exact meaning vague and heavily dependent on specific context.
Additionally, in some cases, those scripts are expected to be prepended with
their byte length. This makes the situation even worse.

The tBTC smart contracts rely on byte-length-prefixed scripts while the
off-chain code leverage the unprefixed version which makes construction of
Bitcoin transactions easier. In order to emphasize the difference, we are
promoting the `bitcoin.Script` type to represent the script version preferred
by the off-chain code. All conversions to/from byte-length-prefixed version
preferred by smart contracts is supposed to be done by chain-specific
code living in the `pkg/chain` package.