Extract chain interface for sweep proposal validation
We extract functions used by the ValidateDepositSweepProposal function
to a separate interface to make usage of this function easier from other
packages. With the changes here to use the ValidateDepositSweepProposal
function the chain implementation passed to the function will require
only two functions to be implemented instead of the whole tbtc.Chain
interface.