Inform about contracts not being deployed to Mainnet
We want to add a note to each generated API doc for Random Beacon that informs
about contracts not being deployed on Mainnet. The note is in style used by
GitBook to highlight important information (read more:
https://docs.gitbook.com/content-creation/blocks/hint#git-sync-representation-in-markdown).
The note is inserted between 5th and 6th line of each generated doc.

Note: This commit contains a temporary change of the triggering conditions for
the publish job. We need to remove this change before merging to `main`.