Define podManagementPolicy: Parallel
The default podManagementPolicy is `OrderedReady`, which waits with
spinning up subsequent replicas untill the previous replica is ready.
In our usecase it's inacurrate, as we want to have three independent
replicas. If the first one is broken, we still want to have a chance to
have another working.

Read more: https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#pod-management-policies