Bitcoin difficulty maintainer for keep-test
Here we update the configuration of bitcoin maintainer. We use the
keep-maintainer template and run the `bitcoinDifficulty` module of the
maintainer connected to mainnet Bitcoin network.

This is expected to be a temporary solution that will be updated to
overcome the limits of LightRelay and Bitcoin testnet integration
mentioned in the README