Remove deprecation notice for --mainnet flag
We planned to remove the `--mainnet` flag, as it seemed unnecessary, as
the client running without any network flag runs for mainnet by default.
But from infrastructure adminstrator perspective the flag may be actually
usefull when running the application on multiple environments. It's much
easier to maintain a common template where the administrator will have
to change value of the flag rather than set the flag for some
environments and remove it for mainnet.