Adding an integration test that hits SubscribeHeaders() concurrently
In this test we spin up 20 goroutines that hit electrum's
client.SubscribeHeaders() concurrently. In the cab1168 commit we
introduced a RW lock on getting the latest block height function.
Without a W lock, the test would've failed with 'concurrent write to websocket connection' panic.
If there's no panic, then the W lock works. 20 goroutine was always
enough to see the panic thrown with only R lock.