Increasing signing attempts to perform all signing computations
It turns out that the current number attempts is not enough and we need
to bump it up so the integration tests computation have enough time to
perform signing. The reason why they started failing is because local
chain has block times much shorter than the real ones. This will give
much shorter timeouts to perform all the signning computation.