Increase maintainer's redemption request amount limit to 15 BTC
The wallet maintainer bot contains a safety net which forces it to not
include requests exceeding 10 BTC in the automatically submitted redemption
proposals. Such requests are processed manually after confirming they are not
part of a broader malicious action. The redemption mechanism proved
to perform well since its release so, we can increase the limit to 15 BTC
and reduce the need of manual intervention.