Wait 2 confirmations before verifying contracts on Etherscan
Without the waiting, the verification of contracts on Etherscan was sometimes
failing with errors like these:
```
NomicLabsHardhatPluginError: Failed to send contract verification request.
Endpoint URL: https://api-sepolia.etherscan.io/api
Reason: The Etherscan API responded that the address
0xAdF78eb37FDE0F3CbF59d11F0DF54b1170a3937b does not have bytecode.
This can happen if the contract was recently deployed and this fact hasn't
propagated to the backend yet.
Try waiting for a minute before verifying your contract. If you are invoking
this from a script, try to wait for five confirmations of your contract
deployment transaction before running the verification subtask.
```
Despite the suggestion to wait 5 confirmation, it seems 2 are enough for us (we
can always increase that number if we find cases where it isn't enough).