Wallet maintainer: Resolve deposit script type at the very end
The script type resolution requires access to transaction data stored on the
Bitcoin blockchain. If the given transaction hasn't actually happened but
was revealed as deposit to the `Bridge`, the maintainer will not be able
to resolve the deposit script type. So far, an error was returned in that
case and the maintainer stopped the entire deposit sweep task. This cannot
happen as the maintainer can be blocked on deposit sweeps for a long time.
Instead, the maintainer should log an error and leave the deposit type
unresolved.