Harden the wallet main UTXO lookup mechanism
The current version of the main UTXO lookup mechanism looks just on the
last five confirmed transactions targeting the wallet public key hash to
determine the plain text main UTXO of the wallet. This mechanism is not ideal
as it doesn't recognize the difference between true wallet transactions and
arbitrary transfers to the wallet public key hash that can be made by
anyone. That means it is enough to craft several arbitrary transfers
to block the main UTXO lookup and prevent the given wallet to perform
actions requested by the wallet coordinator.

To address that problem, we are improving the mechanism to take the
full transaction history into account. To make it efficient, we are
taking just transaction hashes first and fetch full transaction data
only for the latest transactions, where the chance to find the wallet UTXO
is the highest.