Coordination window borderline blocks
Here we add two methods to the `coordinationWindow` type:

The `activePhaseEndBlock` denotes the end block of the active communication
phase and will be used to complete communication in the right moment.

The `endBlock` denotes the end block of the whole window and will be used
to complete the coordination procedure.