Refactoring and cleanup of tasks code
Here we adjust the tasks code to the current requirements. Redemptions and
deposits are now fetched for specific wallet. There is no need
to maintain the possibility to do so for multiple wallets in the same time.
Because of that, we are simplifying the code. By the way, we are adjusting
logging and tests.