Switch Electrum request logs to debug level
Some time ago, we did some major changes in the Electrum client implementation.
Because of that, we added some `INFO` logging around Electrum requests in order
to facilitate potential debugging. As Electrum-changes are now battle-tested,
we can switch those request logs to `DEBUG` level to decrease the amount
of logs produced on `INFO` level.