Require ECDSA DKG result challenger to be an EOA
The `challengeDkgResult` function uses several `try-catch` blocks as part
of its business logic. However, the EVM has a call stack depth limit equal to
1024. A third-party contract can leverage this limitation and force the
`try-catch`-ed calls to revert unconditionally, by using recursion and
letting those calls to be executed at depth 1025. In such a case, the
control flow is passed to the `catch` clauses which may lead to some undesired
side effects like invalidation of a proper DKG result. To address that problem,
we are adding a requirement that `challengeDkgResult` can be only called by
an EOA. This prevents third-party contracts against calling
`challengeDkgResult`.