Include extra data of deposits for sweep fee estimation
Bitcoin scripts of deposits embedding 32-byte extra data are longer (126 bytes)
than the scripts of regular deposits not holding extra data (92 bytes).
That means the actual virtual size of the resulting deposit sweep transaction
is greater than the size assumed by the sweep fee estimator. Here we adjust the
fee estimator to always take the longer script for calculations to avoid
underestimation of the transaction fee. This approach is not ideal but
is safe as it makes sure transactions are not rejected due to too low network
fees.