Adjust `transport` and `authenticatedConnection` components
The `go-libp2p`'s TLS implementation used within the `transport` struct has a
new constructor that accepts the security protocol ID and a list of muxers.
Regarding the first, it's enough to pass the TLS protocol ID exposed by
`go-libp2p`. For the second, it's enough to pass `nil` as the Keep client does
not use Optimized Stream Multiplexer Selection so no mutex is determined
upfront.

Regarding `authenticatedConnection`, there is a need to implement
an additional method `ConnState` that exposes `ConnectionState`. This
can be fetched from the secure connection established by the TLS
encryption layer.

See:
- https://github.com/libp2p/go-libp2p/releases/tag/v0.24.0
- https://github.com/libp2p/specs/blob/50e2cd49c3261eeb71936e4e1b371bd3aa3d7d62/connections/inlined-muxer-negotiation.md