Fix `TestWalletRegistry_getWalletByPublicKeyHash_NotFound` test
This test used a dummy curve point that was not on the curve.
Since Go 1.19, such a behavior leads to a panic. Here we fix
that by using a proper point that is on the curve.

See: https://tip.golang.org/doc/go1.19 > Minor changes to the library > crypto/elliptic
(cherry picked from commit 50b6bd62758695f13877f253ce66ef272e08dc7d)