Optimize `GetBlockHashByNumber` function
So far the `GetBlockHashByNumber` function called the `BlockByNumber`
function of the RPC client. Under the hood, that resulted in an
inefficient `eth_getBlockByNumber` call with `show transaction details` flag
set to `true` which loaded full transaction data of the given block which
is not necessary in the context of `GetBlockHashByNumber`.

Here we improve that behavior by using the `HeaderByNumber` function
of the RPC client. That function calls `eth_getBlockByNumber` with
`show transaction details` flag set to `false` which does not load
transaction data and returns only data specific to the block header.
This is enough to get the hash of the block.