Use longer TTL for pubsub seen messages
Once a message is received and validated, pubsub re-broadcasts it
to other peers and puts it into the seen messages cache. This way,
subsequent arrivals of the same message are not re-broadcasted
unnecessarily. This mechanism is important for the network to avoid
excessive message flooding. The default value used by libp2p is 2 minutes.
However, Keep client messaging sessions are quite time-consuming so,
we use a longer TTL of 5 minutes to reduce flooding risk even further. Worth
noting that this time cannot be too long as the cache may grow excessively and
impact memory consumption.