Single-shot DKG
The currently used DKG retry mechanism based on random exclusion turned
out to be ineffective for a higher number of participating operators.
Such retries have a very small chance for success and produce a lot of
unnecessary network traffic that consumes bandwidth and CPU excessively.

Here we aim to improve the situation. First, we are making DKG a single-shot
process which fails fast if the result cannot be produced during the
first attempt. Second, we are doubling down the announcement period to
maximize participation chance for all selected operators, even those being
at the edge of the network. Last but not least, we are reducing the submission
delay that is preserved between operators attempting to submit the final
result on-chain.

All those changes combined allow to achieve shorter DKG iterations that can
be timed out quicker. This way, we will be able to repeat DKG more often, with
different operator sets.

(cherry picked from commit 7896c7aad56c0c970163120240e17852abe6139d)