Check veto processing delay during redemption proposal generation
The optimistic redemption upgrade introduces a veto mechanism that
enforces a processing delay on each redemption request. The exact
delay value depends on the number of objections raised against the given
redemption request. The `WalletProposalValidator` contract has been
modified to include validation of that delay factor. Here we introduce
the same for the redemption proposal generator. This ensures the generator
issues proposals that conform the on-chain validation rules and coordination
windows are not being wasted.

See: https://github.com/keep-network/tbtc-v2/pull/788