Increase deposit sweep frequency
So far, deposit sweeps were supposed to happen every 48 hours. This frequency
was good for the early days of the system but starts to be insufficient
now, when a significant number of deposits land every day. To optimize
the deposit sweep process, we are increasing its frequency to 12 hours.

(cherry picked from commit 610ef1866e641b953cfa0bc17eaad150b7101537)