Optimize deposit sweep proposal generator
Fetching deposit script type is not needed for sweep proposal generation.
Doing so means making an additional Electrum call for each deposit. This
makes proposal generation longer and more prone to failure.

At the same time, the script type is used only for Maintainer CLI which
is not so important since we have tbtcscan and there is no way to
manually trigger sweeps due to the introduction of RFC 12.

That said, we are removing script type fetch from the process.

(cherry picked from commit 079bb82184284b3b9ff22401fec75e428cb407c6)